requirements.txt
functions-framework==3.*
line-bot-sdk
pygsheets
google.generativeai
mysql-connector-python
yfinance
openpyxl